Ataris Project
This java project is directed so that users/clients can use it in a very simple way. In such a way that users cannot access the database or make changes to it. Simply to access your own information (Register in the system, Login, exit the interface, see the packages we offer and Who we are).